Removed ENV fallback within OpenSearchRequest initialiser.
The problem with this fallback is that the end result is never soemthing we'd actually want.

For example, if the environment was setup with OPENSEARCH_DESCRIPTION_URL = 'http://google.com/opensearch/description.xml' in actuality, you'd never want @base_url to equal this. You'd instead want to run the value through OpenSearchRequest.get_description but doing that at every initialisation also doesn't feel right as we'd be making two requests (potentially) for every search. Instead, we should guide the user to set OpenSearchRequest.base_url = ENV['OPENSEARCH_DESCRIPTION_URL'] as we do in the opensearch prototype at Parliament.